refactor: move orgUnitDistReport component up & use redux components
