+++ src/config/sections.conf.js
@@ -3,7 +3,7 @@ import { ConnectedStandardReport } from '../pages/StandardReport'
 import { ConnectedDataSetReport } from '../pages/DataSetReport'
 import { ConnectedReportingRateSummary } from '../pages/ReportingRateSummary'
 import { ConnectedResource } from '../pages/resource/Resource'
-import { ConnectedOrganisationUnitDistributionReport } from '../pages/organisation-unit-distribution-report/OrganisationUnitDistributionReport'
+import { ConnectedOrganisationUnitDistributionReport } from '../pages/OrganisationUnitDistributionReport'
 
 /* i18n */
 import { i18nKeys } from '../utils/i18n/i18nKeys'
